pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "moats-verify"
version = "0.2.0"
description = "Verify AI-generated text against your documents."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Moats Verify" }]
dependencies = [
  "fastapi>=0.109.0",
  "uvicorn>=0.27.0",
  "python-multipart>=0.0.6",
  "chromadb>=0.4.22",
  "spacy>=3.7.0",
  "PyMuPDF>=1.23.0",
  "python-docx>=1.1.0",
  "ebooklib>=0.18",
  "beautifulsoup4>=4.12.0",
  "flashrank>=0.2.0",
  "httpx>=0.26.0",
  "openai>=1.10.0",
  "anthropic>=0.18.0",
  "click>=8.1.0",
  "pydantic>=2.5.0",
  "python-dotenv>=1.0.0",
]

[project.scripts]
moats = "backend.cli:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["backend*"]
